server/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.3.1
torchvision==0.18.1
numpy==1.26.4
pandas==2.1.4
tqdm==4.66.4
matplotlib==3.7.1
pillow==9.4.0
scikit-learn==1.3.2
Flask==3.0.3    
Flask-Cors==5.0.0